requirements.txt
# Discord関連
discord.py>=2.3.2

# データベース（SQLite + ORM）
sqlalchemy>=1.4

# 実行時の非同期処理に必要なもの（通常Python標準でOK）
# asyncio（標準ライブラリなので書かなくてOK）
feedparser

nest-asyncio>=1.5.6
python-dotenv>=1.0.0
Flask>=2.3.2
requests>=2.31.0